Factor out contents of the catcher plate to a separate class
