+++ osu.Game.Rulesets.Catch/UI/DropAnimation.cs
@@ -1,0 +1,14 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+namespace osu.Game.Rulesets.Catch.UI
+{
+    /// <summary>
+    /// Animation of an object dropped from the catcher plate.
+    /// </summary>
+    public enum DropAnimation
+    {
+        Drop,
+        Explode
+    }
+}
